+++ ficdl/gui/downloader.py
@@ -1,26 +1,45 @@
-import re
-from typing import Union
-
-import os.path
+from dataclasses import dataclass
+from ficdl.scrapers.types import StoryMetadata
+from pathlib import Path
+import queue
+from typing import Optional, Union
 import threading
 import tkinter as tk
 import tkinter.ttk as ttk
 import tkinter.filedialog as filedialog
 import tkinter.messagebox as messagebox
 
-from ..callbacks import ChapterDetails, InitialStoryDetails
-from ..downloader import download_story
+from bs4 import PageElement
+
+from ficdl.callbacks import ChapterDetails, InitialStoryDetails
+from ficdl.config import CONFIG
+from ficdl.downloader import download_story, write_story
+from ficdl.utils import make_path_safe
+from ficdl.writers.types import OutputFormat, WriterOptions
+
+DOWNLOAD_STATE_CHANGED = '<<DownloadStateChanged>>'
+
+@dataclass(eq=False)
+class DownloadFinished:
+    text: list[list[PageElement]]
+    metadata: StoryMetadata
+    cover_path: Optional[Path]
+
+@dataclass(eq=False)
+class SaveFinished:
+    save_path: Path
 
 class Downloader(tk.Frame):
     def __init__(self, master, window):
         super().__init__(master)
         self.window = window
+        self.download_data = queue.SimpleQueue()
+        self.bind(DOWNLOAD_STATE_CHANGED, self.on_download_state_changed)
         self.create_widgets()
 
     def create_widgets(self):
         self.url = tk.StringVar()
         self.cover_path = tk.StringVar()
-        self.output_path = tk.StringVar()
         self.progress_value = tk.IntVar()
 
         row = 0
@@ -36,33 +55,10 @@ class Downloader(tk.Frame):
         ttk.Button(self, text='Browse...', command=self.on_browse_for_cover).grid(row=row, column=2, sticky=tk.W)
 
         row += 1
-        ttk.Label(self, text='Download to: ').grid(row=row, column=0, sticky=tk.W)
-        ttk.Entry(self, textvariable=self.output_path).grid(row=row, column=1, sticky=tk.W)
-        ttk.Button(self, text='Browse...', command=self.on_browse_for_output).grid(row=row, column=2, sticky=tk.W)
-
-        row += 1
-        self.download_button = ttk.Button(self, text='Download', command=self.on_download)
-        self.download_button.grid(row=row, column=0, sticky=tk.W)
-
-        row += 1
-        ttk.Progressbar(self, orient='horizontal', mode='determinate', variable=self.progress_value).grid(row=row, column=0, columnspan=3, sticky=tk.W)
-
-    def on_browse_for_output(self):
-        file = filedialog.asksaveasfilename(
-            parent=self.window,
-            defaultextension='.epub',
-            filetypes=(
-                ('ePub (all eReaders *except* Kindle', '*.epub'),
-                ('MOBI (all Kindles)', '*.mobi'),
-                ('KFX (post-2015 Kindle models)', '*.kfx'),
-            )
-        )
-        
-        if file == '':
-            # Cancel was clicked
-            return
-        
-        self.output_path.set(file)
+        self.download_progress_bar = ttk.Progressbar(self, orient='horizontal', mode='determinate', variable=self.progress_value)
+        self.download_progress_bar.grid(row=row, column=0, columnspan=2, sticky='ew')
+        self.download_button = ttk.Button(self, text='Download...', command=self.on_download)
+        self.download_button.grid(row=row, column=2, sticky=tk.W)
 
     def on_browse_for_cover(self):
         file = filedialog.askopenfilename(
@@ -70,7 +66,7 @@ class Downloader(tk.Frame):
             filetypes=(
                 ('All supported images', '.jpg .jpeg .png'),
                 ('PNG', '.png'),
-                ('JPG', '.jpg .jpeg')
+                ('JPG', '.jpg .jpeg'),
             )
         )
 
@@ -80,41 +76,96 @@ class Downloader(tk.Frame):
         self.cover_path.set(file)
 
     def download_callback(self, details: Union[ChapterDetails, InitialStoryDetails]):
-        if isinstance(details, InitialStoryDetails):
-            self.progress_value.set((1 / details.chapter_count) * 100)
-        elif isinstance(details, ChapterDetails):
-            self.progress_value.set((details.chatper_number / details.chapter_count) * 100)
-        else:
-            raise Exception("jcotton42 forgot to update all the callback stuff")
+        # not run on the UI thread, so use events
+        self.download_data.put(details)
+        self.event_generate(DOWNLOAD_STATE_CHANGED)
+
+    def ask_for_save_location(self, suggest_name: str) -> Optional[Path]:
+        file = filedialog.asksaveasfilename(
+            parent=self.window,
+            defaultextension='.epub',
+            initialfile=suggest_name,
+            filetypes=(
+                ('ePub (all eReaders *except* Kindle)', '.epub'),
+                ('Mobipocket (all Kindles)', '.mobi'),
+                ('PDF', '.pdf'),
+            )
+        )
+
+        if file == '':
+            # Cancel was clicked
+            return None
+
+        return Path(file)
+
+    def on_download_state_changed(self, _event):
+        def save_to_disk(metadata, text, format, output_path, cover_path):
+            write_story(format, WriterOptions(
+                chapter_text=text,
+                metadata=metadata,
+                output_path=output_path,
+                cover_path=cover_path,
+                font_family=CONFIG.default_font_family,
+                font_size=CONFIG.default_font_size,
+                line_height=CONFIG.default_line_height,
+                page_size=CONFIG.default_page_size,
+            ))
+            self.download_data.put(SaveFinished(save_path=output_path))
+            self.event_generate(DOWNLOAD_STATE_CHANGED)
+
+        while not self.download_data.empty():
+            data = self.download_data.get()
+            if isinstance(data, InitialStoryDetails):
+                self.progress_value.set((1 / len(data.metadata.chapter_names)) * 100)
+            elif isinstance(data, ChapterDetails):
+                self.progress_value.set((data.chatper_number / data.chapter_count) * 100)
+            elif isinstance(data, DownloadFinished):
+                suggest_name = make_path_safe(data.metadata.title)
+                save_path = self.ask_for_save_location(suggest_name)
+
+                if save_path is not None:
+                    self.progress_value.set(0)
+                    self.download_progress_bar.configure(mode='indeterminate')
+                    self.download_progress_bar.start()
+                    thread = threading.Thread(target=save_to_disk, args=(
+                        data.metadata,
+                        data.text,
+                        OutputFormat(save_path.suffix.lstrip('.').lower()),
+                        save_path,
+                        data.cover_path,
+                    ))
+                    thread.start()
+                else:
+                    self.progress_value.set(0)
+                    self.download_button.configure(state=tk.NORMAL)
+            elif isinstance(data, SaveFinished):
+                self.download_button.configure(state=tk.NORMAL)
+                self.download_progress_bar.stop()
+                self.progress_value.set(0)
+                self.download_progress_bar.configure(mode='determinate')
+                messagebox.showinfo('Download finished', 'All done.')
+            else:
+                raise Exception("A callback case isn't being handled in the GUI: " + type(data))
 
     def on_download(self):
-        def on_download_inner(url, cover, path, download_button, callback, progress_value):
-            download_story(url, cover, path, None, callback)
-            download_button.configure(state=tk.NORMAL)
-            messagebox.showinfo(title='Download finished', message='All done.')
-            progress_value.set(0)
+        def on_download_inner(url: str, cover: Path):
+            metadata, text = download_story(url, self.download_callback)
+            self.download_data.put(DownloadFinished(text=text, metadata=metadata, cover_path=cover))
+            self.event_generate(DOWNLOAD_STATE_CHANGED)
 
         url = self.url.get().strip()
-        path = self.output_path.get().strip()
         cover = self.cover_path.get().strip()
-
-        if url == '' or path == '':
-            messagebox.showerror(title='Error', message='You must specify both a URL and a save path.')
-            return
 
         if cover == '':
             cover = None
-        elif not os.path.exists(cover):
-            messagebox.showerror(title='Cover does not exist', message='The path you gave for the cover does not exist.')
-            return
+        else:
+            cover = Path(cover)
 
-        _, suffix = os.path.splitext(path)
-
-        if suffix.casefold() != '.epub'.casefold():
-            messagebox.showerror(title='Unsupported format', message='Only ePub books are supported at the moment.')
-            return
+            if not cover.exists():
+                messagebox.showerror(title='Cover does not exist', message='The path you gave for the cover does not exist.')
+                return
 
         self.download_button.configure(state=tk.DISABLED)
 
-        thread = threading.Thread(target=on_download_inner, args=(url, cover, path, self.download_button, self.download_callback, self.progress_value))
+        thread = threading.Thread(target=on_download_inner, args=(url, cover))
         thread.start()